+++ dbus-mqtt-grid/dbus-mqtt-grid.py
@@ -332,7 +332,7 @@ class DbusMqttGridService:
         self._dbusservice.add_path("/ProductId", 0xFFFF)
         self._dbusservice.add_path("/ProductName", productname)
         self._dbusservice.add_path("/CustomName", customname)
-        self._dbusservice.add_path("/FirmwareVersion", "0.1.7-dev (20241008)")
+        self._dbusservice.add_path("/FirmwareVersion", "0.1.7-dev (20241220)")
         # self._dbusservice.add_path('/HardwareVersion', '')
         self._dbusservice.add_path("/Connected", 1)