+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Added: Device type `heatpump`
 * Added: Filter power values out of range
 * Added: Power factor
+* Added: Shelly Gen2+ device support
 * Changed: Broker port missing on reconnect
 * Changed: Default device instance is now `100`
 * Changed: Fixed service not starting sometimes